backend/requirements.txt
flask
flask-cors
transformers
torch
numpy
sentencepiece
python-dotenv
gunicorn
bcrypt
